+++ src/test/java/org/opengis/cite/validation/VerifyRelaxNGValidator.java
@@ -8,11 +8,12 @@ import javax.xml.parsers.DocumentBuilder;
 import javax.xml.parsers.DocumentBuilderFactory;
 import javax.xml.transform.dom.DOMSource;
 import javax.xml.transform.stream.StreamSource;
-import junit.framework.Assert;
 
 import org.apache.xerces.util.XMLCatalogResolver;
+
 import org.junit.After;
 import org.junit.AfterClass;
+import org.junit.Assert;
 import org.junit.Before;
 import org.junit.BeforeClass;
 import org.junit.Rule;